Add variable Print Override environment variables XTRKCADPRINTSCALE=scale_factor XTRKCADPRINTTEXTSCALE=text_factor scale_factor is a floating point number is frequently 1.0 (if needed) text_factor is a floating point number frequently 0.75 (if needed)
These variables are ignored if the print is to a file.

These seem most likely to result from Pango or other parts of the PDF chain to the printer being given bad DPI information which causes the scaling(s) to be inaccurate.

If there is enough use, we can add to the print dialog itself.