Fix for #299 Order of params in HotBar popup menu is backwards
wMenuListAdd has a parameter index which on Windows controls where the menu
entry is positioned.
On Linux, this parameter was ignored and all entries are placed at the beginning
of the list.
In other cases (recent file list, recent messages etc) placing the new entry at
the begining is expected.
For the HotBar, new entries should go at the end.