Add Examples... menu item on the Help menu
This opens the file selection dialog pointing to the examples/ directory
Files loaded this way are treated as ReadOnly, are not added to the recent
files list and do not change the current directory for other file opens.

Also, check if other opened files are ReadOnly.  If so, add an indicator '=' to
the window title and 'Save' invokes the the SaveAs dialog.
File|New sets the ReadOnly flag as well.

Updated the Tip file and Help