Fix bug in saving files -
due to the instance of wFilSelect in using the human readable part of the file filter looking for the sequence ?(*.? before the filetype in the string - if the user did not fill out a suffix.

If this is not present, the code would end up placing an extra ?*? before the filetype. So all filetype descriptions need to end in ?(*.filetype)?.