Add regression info to demos
The addtions look like
REGRESSION START 10 - turnout yard
STRAIGHT 1 0 0 0 0 DEMO 2
	T4 4 36.500000 1.833333 270.000000 0 0.0 0.0 0.0 0.0 0 0 0 0.000000
	E4 41.833333 1.833333 90.000000 0 0.0 0.0 0.0 0.0 0 0 0 0.000000
	END
...
REGRESSION END

The "10" in the 1st line is the PARAMVERSION that was used to generate the masters
The remainder " - turnout  yard" is a description of the test

The track defns are the Expected (master) tracks and are
matched against the Actual tracks at this point in the demo
If there is a difference, the current Actual tracks and the Expected track
are written to 'xtrkcad.regress' in the bin directory on Linux or one up from the
executable dir on Windows.

Each test generates a line on stdout like
REGRESSION CHECK 10 /storage/source/MR/xtrkcad/sf/v5.2/install/bin/../share/xtrkcad/demos/dmease.xtr:92  - cornu easements
If there are diffs a line will be written like
  FAIL: Compare T:1 EP:0 - Pos: Actual [35.417 1.833], Expected [36.500 1.833]
Which means track 1, endPt[0], 'Pos' field differs from the master (in this case by more than 0.01 inches

How to use:
xtrkcad -d regression=2 # for uninterrupts testing
Select a demo, for full testing select the 1st.
Hold down SHIFT and click "Step"
Tests will run continuously until the last test.
You will see each test being run, hopefully with no FAILs

Linux runs clean.  Windows has a few diffs (Pos in dmconn1/2, seg cnt in )
Also dmjntt (Joining to turnable) will ASSERT due to a bad ep index is not run as the 1st demo.
So you run up to that test and then restart on the subsequent tests.