Add new options for SelectMode (Only, Add) and SelectZero (On, Off) to Options->Command.
Change behavior to respect these flags in how Select operates

Ensure that all tracks that will be selected after a click are highlight when moving over the tracks. This means highlighting already selected tracks that will still be selected after the click with or without Ctrl in the two modes and also highlight all connected tracks with Shift.