Selected and Unselected preview highlight tracks look bad on Windows
We had been trying to mimic GTK handling of surfaes using bitmaps.
The plan was to bitblt the Main and Temp bitmaps to the screen on WM_PAINT
This worked fine for simple black lines but looked bad with preview colors.
There was no operator which just copy the non-blank parts of tempD
Instead we used ops which combined the pixels in odd ways.

Now wDrawSetTempMode(TRUE), which is called from TempRedraw, will clear hBmTemp
and then copy hBmMain to it.
WM_PAINT just BitBlt's hBmTemp to the screen.

Old behavious is invoked if hDrawMainBM is TRUE (to be removed)

Removed explicit calls to wDrawClearTemp