AutoSave and Keep xx generations (0-99) of checkpoints between Saves
Change checkpoint names to add Version so that clashes are reduced.

AutoSave prompts for new layout name if not known every
Options->Preferences->AutoSave CheckPoints

Save other than latest checkpoint under the name LayoutNamexx.bkp in the
working directory. Where xx is a generation number that resets to 00
every Options->Preferences->AutoSave CheckPoints.


Notes - If the autosave fails, the checkpoints still exist. If the
latest checkpoint fails, the older checkpoints still exist. If the
filename is changed the checkpoints for the old file still exist.

Setting Checkpoint Freq to 0 also sets AutoSave CheckPoint Freq to 0.
If AutoSave CheckPoint Freq is 0, but Checkpoint Freq is non-zero, the system
still saves 5 checkpoints.

Use of the backups is a manual operation - intended only for use in
extremis - the basic idea is a copy and rename of the backup file is all that
is needed.

Obviously if the save and checkpoints both fail - "Houston, we have a problem!"