Change to individual control of detailed labelsd per track.
All Flexible Track Types can display a table of points and angles as a
description where the points are roughly six inches apart.

This setting enabled in Move Description command. It is not saved in the layout
as is designed for use when printing or to inform the user during construction.

To enable it, you can use 'd' (for details) or use the context menu when hovering
over the desired track.