Generate Turnout paths based on EndPoints and Segments. Not ready for prime time
Enabled by setting the pref option 'misc.EnablePathGeneration' to 1, currently disabled.
Generated paths match paths in shipped param files with a few exceptions which are tagged with '#UseOldPaths'
The debug flag is 'genpaths' with values from 1 to 6.

Hopefully the code is clearer then current code which this code will supplant.

As part of a previous change, options have been added to TURNOUT objects for
PATH_NOCOMBINE for Turntable and Transfer tables and
PATH_OVERRIDE which supresses path generation (for a few cases)
All shipped parameter files have been updated but existing .xtc files have not.
In particular, the code needs to determine if a TURNOUT object is a turn/transfer table.
Also this code doesn't clean up after itself